Refactor kv and nosql
