+++ pkg/keyvalue/nosql/nosql_test.go
@@ -1,17 +1,17 @@
-package dynamodb_test
+package nosql_test
 
 import (
 	"testing"
 
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/maciejgaleja/gosimple/pkg/keyvalue"
-	kvdynamo "github.com/maciejgaleja/gosimple/pkg/keyvalue/dynamodb"
+	kvnosql "github.com/maciejgaleja/gosimple/pkg/keyvalue/nosql"
 	nsqldynamo "github.com/maciejgaleja/gosimple/pkg/nosql/dynamodb"
 	"github.com/maciejgaleja/gosimple/test"
 	"github.com/stretchr/testify/assert"
 )
 
-func TestKeyValue(t *testing.T) {
+func TestKeyValueDynamodb(t *testing.T) {
 	test.DoTestKeyValue(t, func() keyvalue.Store {
 		sess, err := session.NewSessionWithOptions(session.Options{
 			SharedConfigState: session.SharedConfigEnable,
@@ -21,7 +21,7 @@ func TestKeyValue(t *testing.T) {
 
 		nsql := nsqldynamo.NewDynamoDb(sess, "gosimple-test", "key")
 
-		ddb := kvdynamo.NewDynamoDb(nsql, "key", "value")
+		ddb := kvnosql.NewNoSql(nsql, "key", "value")
 		err = ddb.Clear()
 		assert.NoError(t, err)
 		return ddb
